oebb.sh: fixed remote name
The remote that gets created is just called 'origin'.  That was causing
me not to pull in latest updates.  Also, to avoid issues if a branch was
alread created with the 'install' name, this version should create a
nameless branch and try to name it as a second step that is OK to fail.