scripts/layerman: Use git stash pop instead of git stash apply
If we do not use pop then it keeps building stashes unnecessarily
and never removes them. So we should be using git stash pop instead
of git stash apply once the rebase is done on a dirty tree.

This does need a recent git, the one in ubuntu 9.04 won't do

Signed-off-by: Khem Raj <raj.khem@gmail.com>
Signed-off-by: Koen Kooi <koen@dominion.thruhere.net>